backend/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.4
